requirements.txt
attr==0.3.2
colorama==0.4.6
ConfigParser==7.0.0
cryptography==3.4.8
Cython==3.0.11
dl==0.1.0
docutils==0.21.2
filelock==3.15.4
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.3
jnius==1.1.0
keyring==23.5.0
langchain==0.2.12
langchain_chroma==0.1.2
langchain_community==0.2.11
Pillow==10.4.0
protobuf==5.27.3
pyOpenSSL==24.2.1
python-dotenv==1.0.1
redis==5.0.8
Sphinx==8.0.2
thread==2.0.5
tornado==6.4.1
trove_classifiers==2024.7.2
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
